Use the same logic in the mangadetails filter as the reader filter (#543)
* use the same chapter filter for the mangadetails and the reader presenter page

* removed unused class

* move volume detection logic to chapter utils
