+++ app/src/main/java/eu/kanade/tachiyomi/ui/reader/ReaderPresenter.kt
@@ -27,6 +27,7 @@ import eu.kanade.tachiyomi.ui.reader.loader.DownloadPageLoader
 import eu.kanade.tachiyomi.ui.reader.model.ReaderChapter
 import eu.kanade.tachiyomi.ui.reader.model.ReaderPage
 import eu.kanade.tachiyomi.ui.reader.model.ViewerChapters
+import eu.kanade.tachiyomi.util.chapter.ChapterFilter
 import eu.kanade.tachiyomi.util.storage.DiskUtil
 import eu.kanade.tachiyomi.util.system.ImageUtil
 import kotlinx.coroutines.Dispatchers
@@ -53,10 +54,9 @@ class ReaderPresenter(
     private val sourceManager: SourceManager = Injekt.get(),
     private val downloadManager: DownloadManager = Injekt.get(),
     private val coverCache: CoverCache = Injekt.get(),
-    private val preferences: PreferencesHelper = Injekt.get()
+    private val preferences: PreferencesHelper = Injekt.get(),
+    private val chapterFilter: ChapterFilter = Injekt.get()
 ) : BasePresenter<ReaderActivity>() {
-
-    private val readerChapterFilter = ReaderChapterFilter(downloadManager, preferences)
 
     /**
      * The manga loaded in the reader. It can be null when instantiated for a short time.
@@ -101,7 +101,7 @@ class ReaderPresenter(
             ?: error("Requested chapter of id $chapterId not found in chapter list")
 
         val chaptersForReader =
-            readerChapterFilter.filterChapter(dbChapters, manga, selectedChapter)
+            chapterFilter.filterChaptersForReader(dbChapters, manga, selectedChapter)
 
         when (manga.sorting) {
             Manga.SORTING_SOURCE -> ChapterLoadBySource().get(chaptersForReader)
@@ -187,7 +187,7 @@ class ReaderPresenter(
         chapterItems = withContext(Dispatchers.IO) {
             val dbChapters = db.getChapters(manga).executeAsBlocking()
             val list =
-                readerChapterFilter.filterChapter(dbChapters, manga, getCurrentChapter()?.chapter)
+                chapterFilter.filterChaptersForReader(dbChapters, manga, getCurrentChapter()?.chapter)
                 .sortedBy {
                     when (manga.sorting) {
                         Manga.SORTING_NUMBER -> it.chapter_number
